+++ libs/execution/CMakeLists.txt
@@ -30,8 +30,6 @@ set(execution_headers
   hpx/execution/executors/polymorphic_executor.hpp
   hpx/execution/executors/rebind_executor.hpp
   hpx/execution/executors/static_chunk_size.hpp
-  hpx/execution/executors/thread_execution.hpp
-  hpx/execution/executors/thread_execution_information.hpp
   hpx/execution/executors/thread_pool_executor.hpp
   hpx/execution/force_linking.hpp
   hpx/execution/traits/detail/vc/vector_pack_alignment_size.hpp
@@ -72,8 +70,6 @@ set(execution_compat_headers
   hpx/parallel/executors/post_policy_dispatch.hpp
   hpx/parallel/executors/rebind_executor.hpp
   hpx/parallel/executors/static_chunk_size.hpp
-  hpx/parallel/executors/thread_execution.hpp
-  hpx/parallel/executors/thread_execution_information.hpp
   hpx/parallel/executors/thread_pool_executor.hpp
   hpx/parallel/traits/detail/vc/vector_pack_alignment_size.hpp
   hpx/parallel/traits/detail/vc/vector_pack_count_bits.hpp
