+++ libs/include/include/hpx/include/parallel_executor_information.hpp
@@ -10,6 +10,6 @@
 #include <hpx/config.hpp>
 
 #include <hpx/execution/executors/execution_information.hpp>
-#include <hpx/execution/executors/thread_execution_information.hpp>
+#include <hpx/thread_executors/thread_execution_information.hpp>
 
 #endif
