+++ libs/thread_executors/include/hpx/execution/executors/this_thread_executors.hpp
@@ -10,7 +10,7 @@
 #include <hpx/config.hpp>
 
 #if defined(HPX_HAVE_THREAD_EXECUTORS_COMPATIBILITY)
-#include <hpx/execution/executors/thread_execution.hpp>
+#include <hpx/thread_executors/thread_execution.hpp>
 #include <hpx/thread_executors/this_thread_executors.hpp>
 
 namespace hpx { namespace parallel { namespace execution {
