+++ libs/thread_executors/include_compatibility/hpx/parallel/executors/thread_execution.hpp
@@ -5,17 +5,17 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/config.hpp>
-#include <hpx/execution/config/defines.hpp>
-#include <hpx/execution/executors/thread_execution.hpp>
+#include <hpx/thread_executors/config/defines.hpp>
+#include <hpx/thread_executors/thread_execution.hpp>
 
-#if defined(HPX_EXECUTION_HAVE_DEPRECATION_WARNINGS)
+#if defined(HPX_THREAD_EXECUTION_HAVE_DEPRECATION_WARNINGS)
 #if defined(HPX_MSVC)
 #pragma message(                                                               \
     "The header hpx/parallel/executors/thread_execution.hpp is deprecated, \
-    please include hpx/execution/executors/thread_execution.hpp instead")
+    please include hpx/thread_executors/thread_execution.hpp instead")
 #else
 #warning                                                                       \
     "The header hpx/parallel/executors/thread_execution.hpp is deprecated, \
-    please include hpx/execution/executors/thread_execution.hpp instead"
+    please include hpx/thread_executors/thread_execution.hpp instead"
 #endif
 #endif
